+++ teryt_tree/apps.py
@@ -1,7 +1,8 @@
 from django.apps import AppConfig
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 
 class TerytTreeConfig(AppConfig):
     name = "teryt_tree"
     verbose_name = _("Administrative division")
+    default_auto_field = "django.db.models.AutoField"