+++ teryt_tree/rest_framework_ext/viewsets.py
@@ -1,6 +1,6 @@
 import django_filters
 from django.shortcuts import get_object_or_404
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 try:
     from django_filters import rest_framework as filters