+++ tests/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TestConfig(AppConfig):
+    name = "tests"
+    default_auto_field = "django.db.models.AutoField"