+++ tests/dal_ext/test_filters.py
@@ -3,6 +3,7 @@ from django_filters.filterset import FilterSet
 
 from teryt_tree import factories
 from teryt_tree.dal_ext import filters
+import django
 
 from ..models import TestModel
 
@@ -37,8 +38,11 @@ class AreaFilterTestCase(TestCase):
         self.model_1_child = TestModel.objects.create(area=self.jst_1_child)
         self.model_2 = TestModel.objects.create(area=self.jst_2)
 
+        if django.VERSION < (4, 2):
+            self.assertQuerySetEqual = self.assertQuerysetEqual
+
     def assertQsIdsEqual(self, qs, objects):
-        return self.assertQuerysetEqual(
+        return self.assertQuerySetEqual(
             qs, [o.pk for o in objects], lambda o: o.pk, ordered=False
         )
 
@@ -67,8 +71,11 @@ class AreaMultipleFilterTestCase(TestCase):
         self.model_1_child = TestModel.objects.create(area=self.jst_1_child)
         self.model_2 = TestModel.objects.create(area=self.jst_2)
 
+        if django.VERSION < (4, 2):
+            self.assertQuerySetEqual = self.assertQuerysetEqual
+
     def assertQsIdsEqual(self, qs, objects):
-        return self.assertQuerysetEqual(
+        return self.assertQuerySetEqual(
             qs, [o.pk for o in objects], lambda o: o.pk, ordered=False
         )