+++ tests/test_models.py
@@ -11,16 +11,16 @@ from django.test import TestCase
 
 from teryt_tree import models
 
-from django.utils.encoding import force_text
+from django.utils.encoding import force_str
 
 
 class JednostkaAdministracyjnaTestCase(TestCase):
     def test_string_representation(self):
         entry = models.JednostkaAdministracyjna(name="My entry title")
-        self.assertEqual(force_text(entry), entry.name)
+        self.assertEqual(force_str(entry), entry.name)
 
 
 class CategoryTestCase(TestCase):
     def test_string_representation(self):
         entry = models.Category(name="My entry title")
-        self.assertEqual(force_text(entry), entry.name)
+        self.assertEqual(force_str(entry), entry.name)