+++ HISTORY.rst
@@ -3,6 +3,17 @@
 History
 -------
 
+0.18.0 (2023-09-04)
+*******************
+
+* Add support for Django 4.1, 4.2
+* Drop support Django 2.2, 3.0
+* Drop support Python 3.6, 3.7
+* Add support Python 3.10, 3.11
+* Fix deprecation warning for `django.utils.translation.ugettext_lazy`
+* Fix deprecation warning for defines `default_app_config`
+* Fix resource leak in tests
+
 0.17.0 (2020-11-03)
 *******************